+++ frontend/src/components/effects/index.ts
@@ -1,2 +1,4 @@
 export { SpotlightCard } from './SpotlightCard'
 export { FadeIn } from './FadeIn'
+export { default as ClickSpark } from './ClickSpark'
+export { default as DecryptedText } from './DecryptedText'