+++ frontend/src/components/ui/scroll-area.tsx
@@ -39,14 +39,14 @@ function ScrollBar({
       data-orientation={orientation}
       orientation={orientation}
       className={cn(
-        "flex touch-none p-px transition-colors select-none data-horizontal:h-2.5 data-horizontal:flex-col data-horizontal:border-t data-horizontal:border-t-transparent data-vertical:h-full data-vertical:w-2.5 data-vertical:border-l data-vertical:border-l-transparent",
+        "flex touch-none p-px transition-colors select-none data-horizontal:h-2.5 data-horizontal:flex-col data-horizontal:border-t data-horizontal:mt-2 data-horizontal:bg-white/[0.05] data-vertical:h-full data-vertical:w-3 data-vertical:border-l data-vertical:ml-2 data-vertical:bg-white/[0.04]",
         className
       )}
       {...props}
     >
       <ScrollAreaPrimitive.ScrollAreaThumb
         data-slot="scroll-area-thumb"
-        className="relative flex-1 rounded-full bg-border"
+        className="relative flex-1 rounded-full bg-white/30 transition-colors data-[orientation=vertical]:min-h-8 hover:bg-white/50"
       />
     </ScrollAreaPrimitive.ScrollAreaScrollbar>
   )